% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calci.R
\name{add}
\alias{add}
\title{creates a function to add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number}
}
\value{
the sum
}
\description{
creates a function to add two numbers
}
