% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addValue.R
\name{addValue}
\alias{addValue}
\title{Addition}
\usage{
addValue(x, y)
}
\arguments{
\item{x}{First value}

\item{y}{Second Value}
}
\value{
Numeric value
}
\description{
A function that sums two values.
}
\details{
Just play around.
}
\examples{
val <- addValue(1, 3)
}
