% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calci.R
\name{divide}
\alias{divide}
\title{creates a function to divide two numbers}
\usage{
divide(x, y)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number}
}
\value{
the divisor
}
\description{
creates a function to divide two numbers
}
