% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calci.R
\name{multiply}
\alias{multiply}
\title{creates a function to get product of two numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number}
}
\value{
the product
}
\description{
creates a function to get product of two numbers
}
