% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calci.R
\name{subtract}
\alias{subtract}
\title{creates a function to subtract two numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number}
}
\value{
the difference
}
\description{
creates a function to subtract two numbers
}
